requirements.txt
asgiref==3.4.1
async-generator==1.10
certifi==2021.5.30
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.11
django-ckeditor==6.2.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.13.1
gunicorn==20.1.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
matplotlib==3.3.4
mplcursors==0.4
nbclient==0.5.1
nbconvert==6.0.7
nest-asyncio==1.4.2
numpy==1.19.5
Pillow==8.3.2
psycopg2==2.9.3
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.0.1
whitenoise==5.3.0
wincertstore==0.2
